chore: unify old database classes
