+++ node/impl/src/main/java/eu/cloudnetservice/node/impl/database/h2/H2Database.java
@@ -18,8 +18,7 @@ package eu.cloudnetservice.node.impl.database.h2;
 
 import eu.cloudnetservice.driver.document.Document;
 import eu.cloudnetservice.driver.document.DocumentFactory;
-import eu.cloudnetservice.node.impl.database.sql.SQLDatabase;
-import eu.cloudnetservice.node.impl.database.sql.SQLDatabaseProvider;
+import eu.cloudnetservice.node.impl.database.AbstractDatabase;
 import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -32,10 +31,18 @@ import java.util.WeakHashMap;
 import lombok.NonNull;
 import org.jetbrains.annotations.Nullable;
 
-public final class H2Database extends SQLDatabase {
+@Deprecated(forRemoval = true)
+public final class H2Database extends AbstractDatabase {
 
-  public H2Database(@NonNull SQLDatabaseProvider provider, @NonNull String name) {
-    super(provider, name);
+  private static final String TABLE_COLUMN_KEY = "Name";
+  private static final String TABLE_COLUMN_VAL = "Document";
+
+  private final H2DatabaseProvider databaseProvider;
+
+  public H2Database(@NonNull H2DatabaseProvider provider, @NonNull String name) {
+    super(name, provider);
+
+    this.databaseProvider = provider;
 
     // create the table
     provider.executeUpdate(String.format(
